requirements.txt
arrow==1.3.0
blinker==1.7.0
boto3==1.34.11
botocore==1.34.11
click==8.1.7
dominate==2.9.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
urllib3==1.26.18
visitor==0.1.3
Werkzeug==3.0.1
zipp==3.17.0
